Rename: Request와 Response 데이터 클래스 패키지 통합
- response 패키지의 BaseRes를 data 패키지로 이동
- 기존의 data 패키지 내부의 데이터 클래스들 이름 변경 (-Req)
