+++ app/src/main/java/com/example/relay/login/data/LogInLocalReq.kt
@@ -2,8 +2,7 @@ package com.example.relay.login.data
 
 import com.google.gson.annotations.SerializedName
 
-data class SignInLocalData(
-    @SerializedName("name") val name: String,
+data class LogInLocalReq(
     @SerializedName("email") val email: String,
-    @SerializedName("pwd") val pw: String
+    @SerializedName("pwd") val pwd: String
 )
