+++ app/src/main/java/com/example/relay/login/data/SignInLocalReq.kt
@@ -2,7 +2,8 @@ package com.example.relay.login.data
 
 import com.google.gson.annotations.SerializedName
 
-data class LogInLocalData(
+data class SignInLocalReq(
+    @SerializedName("name") val name: String,
     @SerializedName("email") val email: String,
     @SerializedName("pwd") val pwd: String
 )
